Add: enum안에 있던 type들 각각의 파일로 생성
