+++ BankManagerConsoleApp/BankManagerConsoleApp/enum/Menu.swift
@@ -1,0 +1,19 @@
+//
+//  Menu.swift
+//  BankManagerConsoleApp
+//
+//  Created by 리나 on 2021/01/12.
+//
+
+import Foundation
+
+enum Menu: String {
+    case start = "1"
+    case end = "2"
+    
+    static let description = """
+    1: 은행 개점
+    2: 종료
+    입력 :
+    """
+}
